Replace socketserver with a real polling loop
It didnt seem like socketserver had the right hooks for a mainloop,
which is a pity, since it hid the other details.

This is a slash-and-burn conversion to use add a polling loop, which
also gives us the start of a global screen object and immediate support
for multiple clients.